Get candidates

+++ get_candidates.py
@@ -1,10 +1,11 @@
 import argparse
-from pathlib import Path
-import requests
 import json
 import os
 from tqdm import tqdm
 import paths
+import solrqueries as solrq
+from tqdm import tqdm
+import sparqlqueries as sq
 
 def main(args):
 
@@ -30,29 +31,29 @@ def main(args):
 
     dict_candidates = {}
 
-    for mention in mentions_set:
-        #make an http request to wbsearchentities in wikidata
-        #https://www.wikidata.org/w/api.php?action=wbsearchentities&format=json&language=en&type=item&continue=0&search=Madonna&limit=50
-        url=f"https://www.wikidata.org/w/api.php?action=wbsearchentities&format=json&language=en&type=item&continue=0&search={mention}&limit=50"
-        response = requests.get(url)
-        res_obj = response.json()
-        search_res_obj = res_obj.get('search',[])
-        candidate_ids = []
-        for res in search_res_obj:
-            candidate_ids.append(res['id'])
-        dict_candidates[mention] = candidate_ids
-        
-
+    
     #save dictionary to json file
     file_path = paths.DICT_CANDIDATES_PATH
 
-    with open(file_path, 'w') as fp:
-        json.dump(dict_candidates, fp)
+    if os.path.exists(file_path):
+        print("Loading candidates from file, no need to search for them again")
+        with open(file_path) as f:
+            dict_candidates = json.load(f)
+
+    else:
+        for mention in tqdm(mentions_set, "Serching for candidates"):
+            candidate_ids_relevance = set(solrq.solr_search(mention, rows=25))
+            candidate_ids_popularity = set(solrq.solr_search(mention, rows=25, by_popularity=True))
+            candidate_ids = list(candidate_ids_relevance.union(candidate_ids_popularity))
+            dict_candidates[mention] = candidate_ids
+
+        with open(file_path, 'w') as fp:
+            json.dump(dict_candidates, fp)
 
     candidates_set = set()
 
     for candidate_ids in dict_candidates.values():
-        candidates_set.update(candidate_ids[:20])
+        candidates_set.update(candidate_ids)
 
     candidates_set.update(ground_truth_qids)
 
@@ -75,10 +76,9 @@ def main(args):
     for candidate_ids in tqdm(candidates_set):
         #request to wikidata to get the information about the candidate https://www.wikidata.org/w/rest.php/wikibase/v0/entities/items/Q7617093
         candidate_path=folder / f"{candidate_ids}.json"
-        if candidate_path.exists():
-            url=f"https://www.wikidata.org/w/rest.php/wikibase/v0/entities/items/{candidate_ids}"
-            response = requests.get(url)
-            res_obj = response.json()
+        if not candidate_path.exists():
+            candidate_summary = sq.summarize_qid(candidate_ids)
+            res_obj = candidate_summary
             with open(candidate_path, 'w') as fp:
                 json.dump(res_obj, fp)
 
@@ -91,14 +91,14 @@ def main(args):
         if path.is_file() and path.suffix == ".json":
             with open(path) as f:
                 data = json.load(f)
-                if data.get('statements', {}).get('P18'):
-                    for image in data['statements']['P18']:
-                        try:
-                            commons_url = image["value"]["content"]
-                            if not commons_url.startswith("http"):#from a different domain
-                                image_urls.add(commons_url)
-                        except:
-                            pass
+                images=data["images"]
+                for image in images:
+                    try:
+                        commons_url = image
+                        if not commons_url.startswith("http"):#from a different domain
+                            image_urls.add(commons_url)
+                    except:
+                        pass
 
     with open(paths.CANDIDATES_IMAGES_TXT_PATH, 'w') as f:
         for item in image_urls:
